requirements.txt
alembic==1.5.8
allure-python-commons==2.8.13
altgraph==0.17
amqp==2.6.1
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.3.4
attrs==20.3.0
Babel==2.9.1
billiard==3.5.0.5
black==19.10b0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.3.9
colorlog==3.1.4
ConfigArgParse==1.2.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.1
django-celery==3.2.2
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-guardian==1.4.9
django-redis==4.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
ephem==3.7.7.1
Faker==4.0.3
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
flower==0.9.2
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.1.0
httprunner==3.1.4
idna==2.7
importlib-metadata==3.6.0
iniconfig==1.1.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.9.5
jsonpath==0.82
kombu==4.6.11
locust==1.2.3
locustio==0.9.0
loguru==0.4.1
macholib==1.14
Mako==1.1.4
MarkupSafe==1.1.1
more-itertools==8.7.0
msgpack==1.0.0
mysqlclient==2.0.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
psutil==5.7.2
py==1.10.0
pycparser==2.20
pydantic==1.8.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
PyJWT==1.7.1
pymongo==3.11.3
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.11.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==19.0.1
redis==2.10.6
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.8.0
sentry-sdk==0.14.4
six==1.11.0
SQLAlchemy==1.4.10
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.23
vine==1.3.0
wcwidth==0.2.5
Werkzeug==1.0.1
xlwt==1.3.0
zipp==3.4.0
zope.event==4.4
zope.interface==5.1.0
